// frontend/src/api/models/PatientOutputDTO.ts
export interface PatientOutputDTO {
  /**
   *
   * @type {string}
   * @memberof PatientOutputDTO
   */
  id?: string
  /**
   *
   * @type {string}
   * @memberof PatientOutputDTO
   */
  name?: string
  /**
   *
   * @type {Date}
   * @memberof PatientOutputDTO
   */
  createdAt?: Date
  /**
   *
   * @type {Date}
   * @memberof PatientOutputDTO
   */
  updatedAt?: Date
  /**
   *
   * @type {string}
   * @memberof PatientOutputDTO
   */
  phoneNumber?: string
  /**
   *
   * @type {string}
   * @memberof PatientOutputDTO
   */
  address?: string
  /**
   *
   * @type {string}
   * @memberof PatientOutputDTO
   */
  gender?: string
  /**
   *
   * @type {string}
   * @memberof PatientOutputDTO
   */
  description?: string
  /**
   *
   * @type {number}
   * @memberof PatientOutputDTO
   */
  age?: number
  /**
   *
   * @type {string}
   * @memberof PatientOutputDTO
   */
  email?: string
  /**
   *
   * @type {Array<TherapySessionOutputDTO>}
   * @memberof PatientOutputDTO
   */
  therapySessionsOutputDTO?: Array<TherapySessionOutputDTO>
}